requirements.txt
boto3
pyyaml
pandas
matplotlib
numpy
scipy
tqdm
